Enable multiple detached handlers
Previous design could only have one detached handler per application on a hive.
This commit changes that design. Now one can register multiple detached
handlers for an application.

+++ bh/bh_test.go
@@ -48,8 +48,8 @@ func (h *MyHandler) Rcv(m Msg, c RcvContext) error {
 		panic(fmt.Sprintf("Cannot change the key: %v", err))
 	}
 
-	id := c.(*rcvContext).bee.id().Id % uint32(handlers)
-	if id != uint32(hash) {
+	id := c.(*rcvContext).bee.id().Id % uint64(handlers)
+	if id != uint64(hash) {
 		panic(fmt.Sprintf("Invalid message %v received in %v.", m, id))
 	}
